+++ vendored/transducers/transducers/package.lisp
@@ -1,8 +1,9 @@
 (defpackage transducers
   (:use :cl)
   (:shadow #:map #:concatenate #:log #:step #:split
-           #:cons #:count #:first #:last #:max #:min #:find #:string #:vector #:hash-table
-           #:random)
+           #:string #:base-string #:vector #:bit-vector #:hash-table
+           #:cons #:count #:first #:last #:max #:min #:find
+           #:random #:ratio)
   ;; --- Entry Points --- ;;
   (:export #:transduce)
   ;; --- Transducers -- ;;
@@ -17,10 +18,13 @@
            #:sexp
            #:from-csv #:into-csv)
   ;; --- Higher Order Transducers --- ;;
-  (:export #:branch #:inject #:split)
+  (:export #:safe #:branch #:inject #:split)
   ;; --- Reducers -- ;;
-  (:export #:cons #:snoc #:vector #:string #:hash-table
-           #:count #:average #:median #:quantities
+  (:export #:cons #:snoc
+           #:vector #:bit-vector
+           #:string #:base-string
+           #:hash-table
+           #:count #:average #:median #:variance #:quantities #:ratio
            #:any? #:all? #:anyp #:allp #:any #:all
            #:partition
            #:first #:last
@@ -28,7 +32,8 @@
            #:for #:for-each)
   ;; --- Sources --- ;;
   (:export #:ints #:cycle #:repeat #:random #:shuffle
-           #:plist #:reversed)
+           #:plist #:reversed
+           #:chain #:generator #:make-generator #:generator-func #:*done*)
   ;; --- Conditions --- ;;
   (:export #:empty-transduction
            #:imbalanced-plist
@@ -52,3 +57,4 @@
 (defun reduced (item)
   "Wrap a value to signal that reduction has completed."
   (make-reduced :val item))
+