+++ vendored/transducers/transducers/utils.lisp
@@ -5,18 +5,17 @@
         ((symbolp arg) (ensure-function (symbol-function arg)))
         (t (error "Argument is not a function: ~a" arg))))
 
-;; TODO Make this a macro.
-(defun comp (function &rest functions)
+(defmacro comp (function &rest functions)
   "Function composition.
 
 (funcall (comp #'1+ #'length) \"foo\") == (1+ (length \"foo\"))"
-  (reduce (lambda (f g)
-            (let ((f (ensure-function f))
-                  (g (ensure-function g)))
-              (lambda (&rest arguments)
-                (funcall f (apply g arguments)))))
-          functions
-          :initial-value function))
+  (let ((args (gensym "COMP-ARGS-"))
+        (reversed (reverse (cl:cons function functions))))
+    `(lambda (&rest ,args)
+       ,(reduce (lambda (data fn)
+                  `(funcall ,fn ,data))
+                (cdr reversed)
+                :initial-value `(apply ,(car reversed) ,args)))))
 
 #+nil
 (funcall (comp (const 1337) (lambda (n) (* 2 n)) #'1+) 1)
@@ -61,6 +60,8 @@ Borrowed from Clojure, thanks guys."
     (mapc (lambda (k v) (setf (gethash k table) v)) keys vals)
     table))
 
+;; NOTE: 2025-11-08 This is no longer used within `transducers' itself, but I
+;; leave it here as I know I use it secretly in other projects via `::'.
 (declaim (ftype (function (cl:string &key (:separator character)) list) string-split))
 (defun string-split (string &key (separator #\space))
   "Split a string into a list of substrings according to some configurable