+++ vendored/transducers/CHANGELOG.md
@@ -1,13 +1,47 @@
 # Transducers
 
-### Unreleased
+### 1.6.0 (2026-01-15)
 
 #### Added
 
-- Transducer: `sexp` for interpretting a streams of chars or strings as separate
+- Reducer: `ratio` for the percentage of items that pass a predicate.
+- Reducer: `variance`, the result of which can be applied to `sqrt` to find Standard Deviation.
+- Source: `chain` to compose any number of Generators.
+- Source: The `generator` type and its functions have been exposed so that
+  others can write their own Generators.
+
+### 1.5.1 (2025-11-13)
+
+#### Fixed
+
+- `from-csv` now accounts for quotes, which ignore internal commas. Example:
+  `Foo,"10,000",Bar`.
+
+### 1.5.0 (2025-08-16)
+
+#### Added
+
+- Transducer: `sexp` for interpretting streams of chars or strings as separate
   SEXP strings.
+- Transducer: `safe` for handling conditions within individual transducers. 
 - Reducer: `quantities` for counting unique occurrences of streamed items.
 - Reducer: `partition` for splitting the stream results.
+- Reducer: `base-string` for reducing into a `simple-base-string`.
+- Reducer: `bit-vector` for reducing into a `bit-vector`.
+- Can now `transduce` over a `bit-vector` (e.g. `#*0101`).
+- Can now `transduce` specially over `simple-string` and `simple-bit-vector`,
+  but only with certain compilers.
+
+#### Changed
+
+- Reducer: `vector` and `average` are now faster and use less memory.
+- Reducer: `string` is now more efficient and yields the specialized
+  `(simple-array character (*))` type.
+- `comp` is now a macro and uses less memory for long composition chains.
+- All iterations are no longer implicitly wrapped in a `restart-case`. Doing so
+  improves performance (2x in the case of SBCL, 10x in the case of ECL, etc.)
+  and drastically reduces memory usage. As mentioned above, if you know that
+  specific transducers may fail, you can now wrap them in `safe`.
 
 #### Fixed